Fisrt entry!
The idea of the index is to have a simple page asking the user "who dares to enter this hellish domains?". This will give some sort of "start game" mood to the index and make the user to start his/her quest with a simple thing: the name of the character.
To the CSS stylesheet, I chose a "diablo" color pattern, but this will change based on the dungeon the player will be at.

Next entry: modify the style to give the land page more with the user experience and start creating the second page (creating the character).